Add tests for unmanaged mode
These test the behavior when setting `undef` for the `mode` parameter to
the base `selinux` class. The tested behavior is that the module does
not manage the selinux settings at all, leaving the current state the
same.

This is the desired behavior as described in #64, because otherwise
using one of the defined types to manage just a specific selinux rule,
but not manually declaring the `selinux` class and setting the mode
explicitly causes selinux to be disabled. It is confusing (and
undocumented) to use a defined type in a module called `selinux` to set
an selinux rule, and have that result in selinux getting disabled.