Add tests for unmanaged type
These test the behavior when setting `undef` for the `type` parameter to
the base `selinux` class. The tested behavior is that the module does
not manage the selinux settings at all, leaving the current state the
same.

This is the desired behavior as described in #64, because otherwise
using one of the defined types to manage just a specific selinux rule,
but not manually declaring the `selinux` class causes selinux
configuration to potentially change.