Switch to devel package for makefile on RHEL7
Switching to policycoreutils-devel for Redhat/CentOS 7 to provide /usr/share/selinux/devel/Makefile

policycoreutils-python is a dependency, and will be included as well when policycoreutils-devel is installed.

Same effect as using selinux-policy-devel which is the package that actually contains the makefile, but used this to stay consistent.

If still receiving errors after updating module have to delete the .te file that gets copied to /usr/share/selinux folder to get the build process to run again (set to refresh only).