Update and check runtime SELinux status correcty
The #245 issue was addressed by #246, but that introduced a new problem:
when mode param is set to 'disabled', runtime mode never gets changed to
'permissive' anymore.

The original problem was different: it was not wrong parameter to
setenforce, instead, the unless command was wrong, always grepping for
'disabled|disabled'.

This change uses literal modes instead of numeric for both exec *and*
unless command and removes conditinal statement added in #246 to restore
original behaviour. It also adds tests for unless param of the exec in
question.